+++ packages/spotlight/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 The following is a curated list of changes in the Enact spotlight module, newest changes on the top.
 
+## [unreleased]
+
+### Fixed
+
+- `spotlight` to always set focus when the window is activated after the window has been previously blurred
+
+
 ## [2.0.0-beta.2] - 2018-05-07
 
 ### Fixed